Clean up bench
