--- bench/run_2.exs
@@ -1,65 +1,0 @@
-defmodule Bench do
-  def json_xema(path),
-    do:
-      "bench/schema"
-      |> Path.join(path)
-      |> File.read!()
-      |> Jason.decode!()
-      |> JsonXema.new()
-
-  def ex_json_schema(path),
-    do:
-      "bench/schema"
-      |> Path.join(path)
-      |> File.read!()
-      |> Jason.decode!()
-      |> ExJsonSchema.Schema.resolve()
-
-  def json(path),
-    do:
-      "bench/data"
-      |> Path.join(path)
-      |> File.read!()
-      |> Jason.decode!()
-
-  def json_xema_valid?(%{json_xema: schema, json: json}),
-    do: true = JsonXema.valid?(schema, json)
-
-  def ex_json_schema_valid?(%{ex_json_schema: schema, json: json}),
-    do: true = ExJsonSchema.Validator.valid?(schema, json)
-
-  def run do
-    inputs =
-      "bench/schema"
-      |> File.ls!()
-      |> Enum.into(%{}, fn name ->
-        {name,
-         %{
-           json_xema: json_xema(name),
-           ex_json_schema: ex_json_schema(name),
-           json: json(name)
-         }}
-      end)
-
-    Benchee.run(
-      %{
-        "JsonXema" => &json_xema_valid?/1,
-        "ExJsonSchema" => &ex_json_schema_valid?/1
-      },
-      parallel: 4,
-      inputs: inputs,
-      print: [fast_warning: false],
-      formatters: [
-        # &Benchee.Formatters.HTML.output/1,
-        &Benchee.Formatters.Console.output/1
-      ],
-      formatter_options: [
-        html: [
-          file: Path.expand("output/bench.html", __DIR__)
-        ]
-      ]
-    )
-  end
-end
-
-Bench.run()
